Created main method and 2 classes
Main method uses the classes Speaker1 and Speaker2 to run separate
threads to simulate a conversation.  It runs but it does the threads in
sequence because of synchronized methods.  I'm working on changing the
run functions to wait until the other run finishes their side of things.
Still a work in progress though.