Revised Speaker and added Joke class
Speaker class now stores only one name at a time instead of two and
there are new methods.  setRandomName() shuffles the Speaker's name but
if a name is given that is already in use by another Speaker it shuffles
it to the end of the List and chooses from the rest.  The nameInUse
variable is set by the method setNameInUse which takes a String
parameter.  getCurrentName() returns the name in use so that the main
method can use that with another Speaker object to set the nameInUse.
The Joke class currently stores a List<String> that makes up a joke.