Made some changes to Joke and adjusted Speaker and KnockKnock.
Changed method getJokeList to getNextLine since the main method should
not be reading the entire joke list from the Joke object.  Changed it so
that getNextLine would read the next line of the List stored and delete
that line then return the read line to the Speaker object.  Then I
changed the Speaker object method speakJokes so that it uses String
formatting to read the Joke list and place the lines into the output
correctly.  KnockKnock simply had a second speakJokes method call added
to test the speakJokes method.