Major changes including adding JSON
First made changes to Joke so that it holds line1 and line2 which are
Strings parsed from a JSONObject.  Then I added a static method
jsonToJokes that takes a JSONArray and returns a List<Joke> object to be
used by the main method.  Then I added the Joke<List> parameter to the
Speaker constructor instead of having Speaker translate the Jokes into a
List by itself.  Last of all I added the FileReader, BufferedReader, and
try catch blocks to manage the input of the text file Jokes.txt.  Then I
put it all together and had the Speaker use speakJokes twice to test the
output.