Added Thread functionality
Changed Speaker so that it implements Runnable and added a run method
based on speakJokes() but uses a new parameter to determine which
Speaker speaks the first lines and which speaker follows the lead.  Then
added Thread objects to the main method.  The Threads aren't coordinated
but they do run at the same time sometimes in a jumbled order.