Made further changes including extending Speaker into two other classes
Extended Speaker into the Responder and Teller classes since I was
trying to do two different things with Speaker but still used some
common variables to do it.  Teller will be the first speaker's lines and
Responder will be the second speaker's lines.  To help with this added
three protected methods to retrieve Speaker variables for use in Teller
and Responder.